+++ src/main/java/com/numpyninja/lms/services/UserServices.java
@@ -681,66 +681,7 @@ public class UserServices implements UserDetailsService {
     }
     
     
-    public JwtResponseDto forgotPasswordConfirmEmail(EmailDto emailDto) {
-		String userLoginEmail = emailDto.getUserLoginEmailId();
-		Optional<UserLogin> userOptional = userLoginRepository.findByUserLoginEmailIgnoreCase(userLoginEmail);
-		JwtResponseDto forgotPwdDto = new JwtResponseDto();
-		if (userOptional.isPresent()) { // User is present in database
-			UserLogin userLogin = userOptional.get();
-			String userId = userLogin.getUserId();
-            Optional<User> user = userRepository.findById(userId);
-           User userDetails = user.get();
-           
-			if ("active".equalsIgnoreCase(userLogin.getLoginStatus())) {
-				// emailDto.setStatus("active");
-
-				// sending welcome email after user creation
-				try {
-					Map<String, Object> model = new HashMap<>();
-					model.put("firstName", userDetails.getUserFirstName());
-					System.out.println("FirstName"+userDetails.getUserFirstName());
-					model.put("lastName", userDetails.getUserLastName());
-					String token = jwtUtils.generateJwtTokenForgotPwd(userLogin.getUserLoginEmail());
-					String url = createEmailUrlConfirmPwdWithToken(userLogin.getUserLoginEmail(), token);
-					System.out.println("email URL:" + url);
-					model.put("regLink", url);
-
-					String emailMessage = emailSender.sendEmailUsingTemplateForgotPassword(new EmailDetails(userLogin.getUserLoginEmail(), "", "",
-									"Please click on link to generate new password", model));
-					System.out.println(emailMessage);
-					forgotPwdDto.setEmail(userLoginEmail);
-					forgotPwdDto.setToken(token);
-					forgotPwdDto.setStatus("Email sent to your registered email Id");
-
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-
-			} else {
-				forgotPwdDto.setEmail(userLoginEmail);
-				forgotPwdDto.setStatus("login inactive");
-				forgotPwdDto.setToken("null");
-
-			}
-		} else {
-			forgotPwdDto.setEmail(userLoginEmail);
-			forgotPwdDto.setStatus("Invalid Email");
-			forgotPwdDto.setToken("null");
-		}
-
-		return forgotPwdDto;
-	}
-
-	public String createEmailUrlConfirmPwdWithToken(String loginEmail, String token) {
-
-		final String url = UriComponentsBuilder.fromHttpUrl(frontendUrl).path("/reset-password")
-				.queryParam("token", token).toUriString();
-
-		return url;
-
-	}
-
-
+
