+++ src/main/resources/templates/email-templates/ConfirmEmail.ftl
@@ -2,12 +2,9 @@
 <head></head>
 <body>
     <p>Hello ${firstName} ${lastName},</p>
-    <p>Welcome to NumPy Ninja!!</p>
-    <p>You have requested to change your password, You can change your password by clicking below link 
+    <p>You have requested to change your password, You can change your password by clicking given link 
     <a href=${regLink}>click link</a>
-    The link will expire after 15 minutes.</p>
-    <p>Please visit <a href="www.numpyninja.com">Numpy Ninja</a> website
-    to explore more about us!</p>
+    The link will expire after 2 hours.</p>
     <p>Thanks,</p>
     <p>Admin group</p>
     <p>Numpy Ninja</p>
